requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.34
botocore==1.40.34
click==8.2.1
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
